// src/app/shared/enums/end-points.enum.ts
export enum EndPoints {
    Init = 'init',
    Home = 'home',
    About = 'about',
    How = 'how',

    Products = 'product',
    CompanyProducts = 'company',
    Document = 'document',
    News = 'news',
    Offer = 'offer',

    Code = 'code',
    Login = 'auth/login',
    LogOut = 'auth/logout',
    Refresh = 'auth/refresh',
    Register = 'register',
    ChangePassword = 'change-password',
    Recover = 'recover',
    CallMe = 'call-me',
    User = 'user',
    Info = 'info',
    Packages = 'packages',
    Subscribe = 'subscribe',

    Countries = 'countries',
    IntOptions = 'internet-options',
    Cities = 'cities',

    Ads = 'ads',
    AdsClose = 'ads/close',

    Order = 'order',
    ConfirmOrder = 'confirm-order',
    CheckSmsCode = 'check-sms-code',
    CheckCaptchaSubscriber = 'check-captcha-subscriber',
}